[improvement][Headless] Simplify the QueryService interface, optimize Query permissions, and add integration testing.
