+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/annotation/S2DataPermission.java
@@ -9,6 +9,6 @@ import java.lang.annotation.Documented;
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface S2SQLDataPermission {
+public @interface S2DataPermission {
 
 }
