--- headless/server/src/main/java/com/tencent/supersonic/headless/server/annotation/StructDataPermission.java
@@ -1,12 +1,0 @@
-package com.tencent.supersonic.headless.server.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.PARAMETER, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface StructDataPermission {
-
-}
