+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/rest/QueryController.java
@@ -8,7 +8,6 @@ import com.tencent.supersonic.headless.api.pojo.request.BatchDownloadReq;
 import com.tencent.supersonic.headless.api.pojo.request.DownloadStructReq;
 import com.tencent.supersonic.headless.api.pojo.request.ExplainSqlReq;
 import com.tencent.supersonic.headless.api.pojo.request.ItemUseReq;
-import com.tencent.supersonic.headless.api.pojo.request.ParseSqlReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryDimValueReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryItemReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryMultiStructReq;
@@ -18,22 +17,18 @@ import com.tencent.supersonic.headless.api.pojo.response.ExplainResp;
 import com.tencent.supersonic.headless.api.pojo.response.ItemQueryResultResp;
 import com.tencent.supersonic.headless.api.pojo.response.ItemUseResp;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticQueryResp;
-import com.tencent.supersonic.headless.api.pojo.response.SqlParserResp;
-import com.tencent.supersonic.headless.core.pojo.QueryStatement;
 import com.tencent.supersonic.headless.server.service.DownloadService;
 import com.tencent.supersonic.headless.server.service.QueryService;
+import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.validation.Valid;
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/semantic/query")
@@ -51,7 +46,7 @@ public class QueryController {
             HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         User user = UserHolder.findUser(request, response);
-        return queryService.queryBySql(querySQLReq, user);
+        return queryService.queryByReq(querySQLReq, user);
     }
 
     @PostMapping("/struct")
@@ -60,7 +55,7 @@ public class QueryController {
             HttpServletResponse response) throws Exception {
         User user = UserHolder.findUser(request, response);
         QuerySqlReq querySqlReq = queryStructReq.convert(queryStructReq, true);
-        return queryService.queryBySql(querySqlReq, user);
+        return queryService.queryByReq(querySqlReq, user);
     }
 
     @PostMapping("/queryMetricDataById")
@@ -85,19 +80,6 @@ public class QueryController {
         downloadService.batchDownload(batchDownloadReq, user, response);
     }
 
-    @PostMapping("/queryStatement")
-    public SemanticQueryResp queryStatement(@RequestBody QueryStatement queryStatement) throws Exception {
-        return queryService.queryByQueryStatement(queryStatement);
-    }
-
-    @PostMapping("/struct/parse")
-    public SqlParserResp parseByStruct(@RequestBody ParseSqlReq parseSqlReq) throws Exception {
-        QueryStatement queryStatement = queryService.explain(parseSqlReq);
-        SqlParserResp sqlParserResp = new SqlParserResp();
-        BeanUtils.copyProperties(queryStatement, sqlParserResp);
-        return sqlParserResp;
-    }
-
     /**
      * queryByMultiStruct
      */
@@ -106,7 +88,7 @@ public class QueryController {
             HttpServletRequest request,
             HttpServletResponse response) throws Exception {
         User user = UserHolder.findUser(request, response);
-        return queryService.queryByMultiStruct(queryMultiStructReq, user);
+        return queryService.queryByReq(queryMultiStructReq, user);
     }
 
     /**
@@ -132,23 +114,19 @@ public class QueryController {
     public <T> ExplainResp explain(@RequestBody ExplainSqlReq<T> explainSqlReq,
             HttpServletRequest request,
             HttpServletResponse response) throws Exception {
-
         User user = UserHolder.findUser(request, response);
         String queryReqJson = JsonUtil.toString(explainSqlReq.getQueryReq());
-        QueryType queryTypeEnum = explainSqlReq.getQueryTypeEnum();
 
-        if (QueryType.SQL.equals(queryTypeEnum)) {
-            QuerySqlReq querySQLReq = JsonUtil.toObject(queryReqJson, QuerySqlReq.class);
+        if (QueryType.SQL.equals(explainSqlReq.getQueryTypeEnum())) {
             ExplainSqlReq<QuerySqlReq> explainSqlReqNew = ExplainSqlReq.<QuerySqlReq>builder()
-                    .queryReq(querySQLReq)
-                    .queryTypeEnum(queryTypeEnum).build();
+                    .queryReq(JsonUtil.toObject(queryReqJson, QuerySqlReq.class))
+                    .queryTypeEnum(explainSqlReq.getQueryTypeEnum()).build();
             return queryService.explain(explainSqlReqNew, user);
         }
-        if (QueryType.STRUCT.equals(queryTypeEnum)) {
-            QueryStructReq queryStructReq = JsonUtil.toObject(queryReqJson, QueryStructReq.class);
+        if (QueryType.STRUCT.equals(explainSqlReq.getQueryTypeEnum())) {
             ExplainSqlReq<QueryStructReq> explainSqlReqNew = ExplainSqlReq.<QueryStructReq>builder()
-                    .queryReq(queryStructReq)
-                    .queryTypeEnum(queryTypeEnum).build();
+                    .queryReq(JsonUtil.toObject(queryReqJson, QueryStructReq.class))
+                    .queryTypeEnum(explainSqlReq.getQueryTypeEnum()).build();
             return queryService.explain(explainSqlReqNew, user);
         }
         return null;
