--- launchers/standalone/src/test/java/com/tencent/supersonic/benchmark/CSpider.java
@@ -1,10 +1,0 @@
-package com.tencent.supersonic.benchmark;
-
-import org.junit.Test;
-
-public class CSpider {
-    @Test
-    public void case1(){
-
-    }
-}
