+++ launchers/standalone/src/test/java/com/tencent/supersonic/chat/integration/mapper/MapperTest.java
@@ -1,5 +1,9 @@
-package com.tencent.supersonic.integration.mapper;
+package com.tencent.supersonic.chat.integration.mapper;
 
+import static com.tencent.supersonic.common.pojo.enums.AggregateTypeEnum.NONE;
+
+import com.tencent.supersonic.chat.integration.BaseTest;
+import com.tencent.supersonic.chat.integration.util.DataUtils;
 import com.tencent.supersonic.chat.api.pojo.SchemaElement;
 import com.tencent.supersonic.chat.api.pojo.SemanticParseInfo;
 import com.tencent.supersonic.chat.api.pojo.request.QueryFilter;
@@ -9,13 +13,9 @@ import com.tencent.supersonic.chat.core.query.rule.metric.MetricTagQuery;
 import com.tencent.supersonic.common.pojo.DateConf;
 import com.tencent.supersonic.common.pojo.enums.FilterOperatorEnum;
 import com.tencent.supersonic.common.pojo.enums.QueryType;
-import com.tencent.supersonic.integration.BaseQueryTest;
-import com.tencent.supersonic.util.DataUtils;
 import org.junit.Test;
 
-import static com.tencent.supersonic.common.pojo.enums.AggregateTypeEnum.NONE;
-
-public class MapperTest extends BaseQueryTest {
+public class MapperTest extends BaseTest {
 
     @Test
     public void hanlp() throws Exception {
