+++ launchers/standalone/src/test/java/com/tencent/supersonic/chat/integration/plugin/PluginRecognizeTest.java
@@ -1,5 +1,6 @@
-package com.tencent.supersonic.integration.plugin;
+package com.tencent.supersonic.chat.integration.plugin;
 
+import com.tencent.supersonic.chat.integration.util.DataUtils;
 import com.tencent.supersonic.chat.api.pojo.request.ExecuteQueryReq;
 import com.tencent.supersonic.chat.api.pojo.request.QueryFilter;
 import com.tencent.supersonic.chat.api.pojo.request.QueryFilters;
@@ -10,8 +11,7 @@ import com.tencent.supersonic.chat.core.plugin.PluginManager;
 import com.tencent.supersonic.chat.server.service.AgentService;
 import com.tencent.supersonic.chat.server.service.QueryService;
 import com.tencent.supersonic.common.config.EmbeddingConfig;
-import com.tencent.supersonic.integration.MockConfiguration;
-import com.tencent.supersonic.util.DataUtils;
+import com.tencent.supersonic.chat.integration.MockConfiguration;
 import org.junit.Assert;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
